fix: rename pubsub file
